requirements.txt
asgiref==3.2.3
Django==3.0.2
django-rest-framework==0.1.0
djangorestframework==3.11.0
pigpio==1.45
psutil==5.6.7
pytz==2019.3
sqlparse==0.3.0
typing-extensions==3.7.4.1
